added todos, and external data module working
