+++ public/models/dummy_data.js
@@ -1,0 +1,20 @@
+
+
+//static data
+var myData = [
+    { "first":"jeff", "last":"wu", "title":"creative technlogist" },
+    { "first":"shirley", "last":"yang", "title":"teacher" },
+    { "first":"wynbert", "last":"gan", "title":"IT director" },
+    { "first":"wayne", "last":"dong", "title":"IT specialist" }
+];
+
+//static quotes data
+var quotes = [
+    { author : 'Audrey Hepburn', text : "Nothing is impossible, the word itself says 'I'm possible'!"},
+    { author : 'Walt Disney', text : "You may not realize it when it happens, but a kick in the teeth may be the best thing in the world for you"},
+    { author : 'Unknown', text : "Even the greatest was once a beginner. Don't be afraid to take that first step."},
+    { author : 'Neale Donald Walsch', text : "You are afraid to die, and you're afraid to live. What a way to exist."}
+];
+
+module.exports.myData = myData;
+module.exports.quotes = quotes;
